Continuation of dose-response controller splitting
