+++ src/main/java/be/ugent/maf/cellmissy/gui/controller/analysis/doseresponse/area/AreaDRNormalizedController.java
@@ -3,11 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse;
+package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse.area;
 
 import be.ugent.maf.cellmissy.entity.PlateCondition;
 import be.ugent.maf.cellmissy.entity.Treatment;
 import be.ugent.maf.cellmissy.entity.result.doseresponse.AreaDoseResponseAnalysisGroup;
+import be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse.DRNormalizedController;
 import be.ugent.maf.cellmissy.gui.experiment.analysis.doseresponse.DRNormalizedPlotPanel;
 import be.ugent.maf.cellmissy.gui.view.table.model.NonEditableTableModel;
 import be.ugent.maf.cellmissy.utils.AnalysisUtils;
@@ -18,6 +19,7 @@ import java.awt.event.ActionListener;
 import java.awt.event.ItemEvent;
 import java.awt.event.ItemListener;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -32,23 +34,16 @@ import org.springframework.stereotype.Controller;
  * @author Gwendolien
  */
 @Controller("areaDRNormalizedController")
-public class AreaDRNormalizedController implements DRNormalizedController {
+public class AreaDRNormalizedController extends DRNormalizedController {
 
     private static final org.apache.log4j.Logger LOG = org.apache.log4j.Logger.getLogger(AreaDRNormalizedController.class);
 
     //model
-    private Double bottomConstrainValue;
-    private Double topConstrainValue;
-    private NonEditableTableModel tableModel;
     private LinkedHashMap<Double, List<Double>> dataToFit;
-    //view
-    private DRNormalizedPlotPanel dRNormalizedPlotPanel;
-    private ChartPanel normalizedChartPanel;
+    //view: in super class
     // parent controller
     @Autowired
-    private DoseResponseController doseResponseController;
-    // services
-    private GridBagConstraints gridBagConstraints;
+    private AreaDoseResponseController doseResponseController;
 
     /**
      * Initialize controller
@@ -64,32 +59,8 @@ public class AreaDRNormalizedController implements DRNormalizedController {
      *
      * @return
      */
-    public DRNormalizedPlotPanel getDRNormalizedPlotPanel() {
-        return dRNormalizedPlotPanel;
-    }
-
-    public NonEditableTableModel getTableModel() {
-        return tableModel;
-    }
-
-    private void setTableModel(NonEditableTableModel tableModel) {
-        this.tableModel = tableModel;
-    }
-
-    public ChartPanel getNormalizedChartPanel() {
-        return normalizedChartPanel;
-    }
-
     public LinkedHashMap<Double, List<Double>> getDataToFit() {
         return dataToFit;
-    }
-
-    public Double getBottomConstrainValue() {
-        return bottomConstrainValue;
-    }
-
-    public Double getTopConstrainValue() {
-        return topConstrainValue;
     }
 
     /**
@@ -101,8 +72,8 @@ public class AreaDRNormalizedController implements DRNormalizedController {
         dRNormalizedPlotPanel.getBottomComboBox().setSelectedIndex(0);
         dRNormalizedPlotPanel.getTopComboBox().setSelectedIndex(0);
         //set initial parameters
-        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMeans(doseResponseController.getdRAnalysisGroup()))).toString());
-        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup()))).toString());
+        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMeans(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
+        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
         //LogTransform concentrations and perform initial normalization (mean values)
         dataToFit = prepareFittingData(doseResponseController.getdRAnalysisGroup());
         //create and set the table model for the top panel table (dependent on normalization)
@@ -136,11 +107,11 @@ public class AreaDRNormalizedController implements DRNormalizedController {
                 switch (value) {
                     case "Smallest Mean Value":
                         dRNormalizedPlotPanel.getBottomTextField().setEditable(false);
-                        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMeans(doseResponseController.getdRAnalysisGroup()))).toString());
+                        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMeans(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
                         break;
                     case "Smallest Median Value":
                         dRNormalizedPlotPanel.getBottomTextField().setEditable(false);
-                        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMedians(doseResponseController.getdRAnalysisGroup()))).toString());
+                        dRNormalizedPlotPanel.getBottomTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.min(computeMedians(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
                         break;
                     case "Other Value":
                         dRNormalizedPlotPanel.getBottomTextField().setText("");
@@ -162,11 +133,11 @@ public class AreaDRNormalizedController implements DRNormalizedController {
                 switch (choice) {
                     case "Largest Mean Value":
                         dRNormalizedPlotPanel.getTopTextField().setEditable(false);
-                        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup()))).toString());
+                        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
                         break;
                     case "Largest Median Value":
                         dRNormalizedPlotPanel.getTopTextField().setEditable(false);
-                        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup()))).toString());
+                        dRNormalizedPlotPanel.getTopTextField().setText(AnalysisUtils.roundTwoDecimals(Collections.max(computeMeans(doseResponseController.getdRAnalysisGroup().getVelocitiesMap().values()))).toString());
                         break;
                     case "Other Value":
                         dRNormalizedPlotPanel.getTopTextField().setText("");
@@ -235,49 +206,6 @@ public class AreaDRNormalizedController implements DRNormalizedController {
      * Private methods
      */
     /**
-     * Compute mean values of every condition in the dose response analysis
-     * group
-     */
-    private List<Double> computeMeans(AreaDoseResponseAnalysisGroup doseResponseAnalysisGroup) {
-        List<Double> allMeans = new ArrayList();
-        for (List<Double> velocities : doseResponseAnalysisGroup.getVelocitiesMap().values()) {
-            List<Double> tempData = new ArrayList();
-            for (Double velocity : velocities) {
-                if (velocity != null) {
-                    tempData.add(velocity);
-                }
-            }
-            double[] data = new double[tempData.size()];
-            for (int i = 0; i < data.length; i++) {
-                data[i] = tempData.get(i);
-            }
-            allMeans.add(AnalysisUtils.computeMean(data));
-        }
-        return allMeans;
-    }
-
-    /**
-     * Compute median values of every condition in DR analysis group
-     */
-    private List<Double> computeMedians(AreaDoseResponseAnalysisGroup doseResponseAnalysisGroup) {
-        List<Double> allMedians = new ArrayList();
-        for (List<Double> velocities : doseResponseAnalysisGroup.getVelocitiesMap().values()) {
-            List<Double> tempData = new ArrayList();
-            for (Double velocity : velocities) {
-                if (velocity != null) {
-                    tempData.add(velocity);
-                }
-            }
-            double[] data = new double[tempData.size()];
-            for (int i = 0; i < data.length; i++) {
-                data[i] = tempData.get(i);
-            }
-            allMedians.add(AnalysisUtils.computeMedian(data));
-        }
-        return allMedians;
-    }
-
-    /**
      * Prepare data for fitting starting from the analysis group.
      *
      * @param dRAnalysisGroup
@@ -288,8 +216,8 @@ public class AreaDRNormalizedController implements DRNormalizedController {
         LinkedHashMap<Double, List<Double>> result = new LinkedHashMap<>();
 
         //!! control concentrations (10 * lower than lowest treatment conc) also need to be added
-        List<List<Double>> allVelocities = new ArrayList<List<Double>>();
-        List< Double> allLogConcentrations = new ArrayList<Double>();
+        List<List<Double>> allVelocities = new ArrayList<>();
+        List<Double> allLogConcentrations = new ArrayList<>();
 
         //put concentrations of treatment to analyze (control not included!) in list
         LinkedHashMap<Double, String> nestedMap = dRAnalysisGroup.getConcentrationsMap().get(dRAnalysisGroup.getTreatmentToAnalyse());
@@ -326,84 +254,6 @@ public class AreaDRNormalizedController implements DRNormalizedController {
         }
 
         return result;
-    }
-
-    /**
-     * Perform normalization on velocities
-     */
-    private Double normalize(Double velocity) {
-        //check which values will become 0% and 100%
-        Double bottomNormalize = Double.parseDouble(dRNormalizedPlotPanel.getBottomTextField().getText());
-        Double topNormalize = Double.parseDouble(dRNormalizedPlotPanel.getTopTextField().getText());
-
-        if (velocity == null) {
-            return velocity;
-        } else if (velocity.isNaN()) {
-            return velocity;
-        }
-        return 100 * (velocity - bottomNormalize) / (topNormalize - bottomNormalize);
-    }
-
-    /**
-     * Create the table model for the top panel table. Table contains
-     * log-transformed concentration and normalized slopes per condition
-     *
-     * @param dataToFit
-     * @return the model
-     */
-    private NonEditableTableModel createTableModel(LinkedHashMap<Double, List<Double>> dataToFit) {
-        int maxReplicates = 0;
-        for (Map.Entry<Double, List<Double>> entry : dataToFit.entrySet()) {
-            int replicates = entry.getValue().size();
-            if (replicates > maxReplicates) {
-                maxReplicates = replicates;
-            }
-        }
-        Object[][] data = new Object[dataToFit.size()][maxReplicates + 1];
-
-        int rowIndex = 0;
-        for (Map.Entry<Double, List<Double>> entry : dataToFit.entrySet()) {
-            //log concentration is put on 1st column
-            data[rowIndex][0] = AnalysisUtils.roundThreeDecimals(entry.getKey());
-
-            for (int columnIndex = 1; columnIndex < entry.getValue().size() + 1; columnIndex++) {
-                Double slope = entry.getValue().get(columnIndex - 1);
-                if (slope != null && !slope.isNaN()) {
-                    // round to three decimals slopes and coefficients
-                    slope = AnalysisUtils.roundThreeDecimals(entry.getValue().get(columnIndex - 1));
-                    // show in table slope + (coefficient)
-                    data[rowIndex][columnIndex] = slope;
-                } else if (slope == null) {
-                    data[rowIndex][columnIndex] = "excluded";
-                } else if (slope.isNaN()) {
-                    data[rowIndex][columnIndex] = "NaN";
-                }
-            }
-            rowIndex++;
-        }
-        // array of column names for table model
-        String[] columnNames = new String[data[0].length];
-        columnNames[0] = "Log-concentration";
-        for (int x = 1; x < columnNames.length; x++) {
-            columnNames[x] = "Repl " + (x);
-        }
-
-        NonEditableTableModel nonEditableTableModel = new NonEditableTableModel();
-        nonEditableTableModel.setDataVector(data, columnNames);
-        return nonEditableTableModel;
-    }
-
-    /**
-     * Give information on how normalization was performed for the PDF report.
-     * @return 
-     */
-    public String getNormalizationInfo() {
-        Double bottomNormalize = Double.parseDouble(dRNormalizedPlotPanel.getBottomTextField().getText());
-        Double topNormalize = Double.parseDouble(dRNormalizedPlotPanel.getTopTextField().getText());
-        String result = "NORMALIZATION: ";
-        String zero = " 0% = " + bottomNormalize + " (" + (String) dRNormalizedPlotPanel.getBottomComboBox().getSelectedItem() + ")  ";
-        String hundred = " 100% = " + topNormalize + " (" + (String) dRNormalizedPlotPanel.getTopComboBox().getSelectedItem() + ")  ";
-        return result + zero + hundred;
     }
 
 }
