+++ src/main/java/be/ugent/maf/cellmissy/gui/controller/analysis/doseresponse/generic/GenericDRInitialController.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse;
+package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse.generic;
 
+import be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse.DRInitialController;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 
 /**
@@ -12,7 +14,13 @@ import org.springframework.stereotype.Controller;
  * @author CompOmics Gwen
  */
 @Controller("genericDRInitialController")
-public class GenericDRInitialController {
+public class GenericDRInitialController extends DRInitialController {
     
     private static final org.apache.log4j.Logger LOG = org.apache.log4j.Logger.getLogger(GenericDRInitialController.class);
+    
+    //model
+    //view
+    //parent controller
+    @Autowired
+    private ;
 }
