+++ src/main/java/be/ugent/maf/cellmissy/gui/controller/analysis/doseresponse/generic/GenericDRInputController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse;
+package be.ugent.maf.cellmissy.gui.controller.analysis.doseresponse.generic;
 
 import org.springframework.stereotype.Controller;
